refactor(hickey): split diagnostic info sections
