+++ packages/client/src/DiagnosticInfo.tsx
@@ -1,159 +1,25 @@
 /** Diagnostic Info — runtime state dump for support tickets and
  *  self-diagnosis. Opened from command palette → Debug → Diagnostic
- *  info. Content split into `<DiagnosticInfoContent/>` so a future
- *  always-visible dev inspector can reuse it without the modal chrome. */
+ *  info. */
 
 import Dialog from "@corvu/dialog";
-import type { ServerDiagnostics, TerminalId } from "kolu-common";
-import {
-  type Component,
-  createMemo,
-  createResource,
-  For,
-  Show,
-} from "solid-js";
+import type { TerminalId } from "kolu-common";
+import type { Component } from "solid-js";
 import { toast } from "solid-sonner";
-import { client, serverProcessId, wsStatus } from "./rpc/rpc";
-import { getTerminalRefs } from "./terminal/terminalRefs";
-import { getDiagnostics } from "./terminal/useTerminalDiagnostics";
-import { webglLifecycleSnapshot } from "./terminal/webglTracker";
+import BrowserDiagnosticsSection from "./diagnostics/BrowserDiagnosticsSection";
+import ServerDiagnosticsSection from "./diagnostics/ServerDiagnosticsSection";
+import { useDiagnosticSnapshot } from "./diagnostics/useDiagnosticSnapshot";
+import XtermDiagnosticsSection from "./diagnostics/XtermDiagnosticsSection";
 import ModalDialog, { refocusTerminal } from "./ui/ModalDialog";
-import Row from "./ui/Row";
-import Section from "./ui/Section";
-import { isMobile } from "./useMobile";
-
-/** WebGL2 support detection creates a throwaway canvas + WebGL context
- *  that lingers on a detached node until GC. Compute once at module load
- *  so re-opening this dialog doesn't burn one context per open — the exact
- *  zombie-context pattern this dialog exists to diagnose (#591). */
-const WEBGL2_SUPPORTED = (() => {
-  const canvas = document.createElement("canvas");
-  return !!canvas.getContext("webgl2");
-})();
-
-/** One-shot browser facts read at first render. Stable for the session,
- *  so no reactive source needed — keeps this module's dependency surface
- *  small. */
-function browserFacts() {
-  return {
-    userAgent: navigator.userAgent,
-    webgl2Supported: WEBGL2_SUPPORTED,
-    crossOriginIsolated: self.crossOriginIsolated,
-    devicePixelRatio: window.devicePixelRatio,
-    xtermVersion: __XTERM_VERSION__,
-  };
-}
-
-/** Single source of truth for byte-count display across the dialog.
- *  `bytesToMB` returns a number (used by the `jsHeap` snapshot shape, which
- *  callers may parse programmatically). `formatMB` returns a display string
- *  and drops to KB below 100 KB — a fresh 80×24 buffer is ~23 KB, and
- *  "0.0 MB" obscures more than it communicates. Every byte render in this
- *  module goes through these two; evolving the granularity is one edit. */
-function bytesToMB(bytes: number): number {
-  return Math.round((bytes / 1_048_576) * 10) / 10;
-}
-function formatMB(bytes: number): string {
-  if (bytes < 100_000) return `${Math.round(bytes / 1024)} KB`;
-  return `${bytesToMB(bytes).toFixed(1)} MB`;
-}
-
-function formatDuration(ms: number): string {
-  const seconds = Math.floor(ms / 1000);
-  if (seconds < 60) return `${seconds}s`;
-  const minutes = Math.floor(seconds / 60);
-  if (minutes < 60) return `${minutes}m ${seconds % 60}s`;
-  const hours = Math.floor(minutes / 60);
-  return `${hours}h ${minutes % 60}m`;
-}
-
-function formatDetails(
-  details: ServerDiagnostics["trackedResources"][number]["details"],
-): string | null {
-  const entries = Object.entries(details);
-  if (entries.length === 0) return null;
-  return entries.map(([key, value]) => `${key}=${String(value)}`).join(" · ");
-}
-
-/** `performance.memory` is Chromium-only and missing from the DOM type
- *  definitions — isolate the narrow cast here so the snapshot memo stays
- *  free of it. Returns null on non-Chromium browsers. */
-function readJsHeap(): {
-  usedMB: number;
-  totalMB: number;
-  limitMB: number;
-} | null {
-  const mem = (
-    performance as {
-      memory?: {
-        usedJSHeapSize: number;
-        totalJSHeapSize: number;
-        jsHeapSizeLimit: number;
-      };
-    }
-  ).memory;
-  if (!mem) return null;
-  return {
-    usedMB: bytesToMB(mem.usedJSHeapSize),
-    totalMB: bytesToMB(mem.totalJSHeapSize),
-    limitMB: bytesToMB(mem.jsHeapSizeLimit),
-  };
-}
 
 const DiagnosticInfoContent: Component<{
   open: boolean;
   activeId: TerminalId | null;
 }> = (props) => {
-  const browser = browserFacts();
-  const [serverDiagnostics] = createResource(
-    () => (props.open ? serverProcessId() : undefined),
-    () => client.server.diagnostics(),
-  );
-
-  const snapshot = createMemo(() => {
-    const webgl = webglLifecycleSnapshot();
-    return {
-      browser,
-      session: {
-        viewport: isMobile() ? "mobile" : "canvas",
-        wsStatus: wsStatus(),
-        serverProcessId: serverProcessId(),
-        activeId: props.activeId,
-        terminalCount: getDiagnostics().length,
-        jsHeap: readJsHeap(),
-        domNodes: document.getElementsByTagName("*").length,
-        canvases: webgl.totalDomCanvases,
-      },
-      server: serverDiagnostics() ?? null,
-      terminals: getDiagnostics().map((d) => {
-        const refs = getTerminalRefs(d.id);
-        const bufferLen = refs?.xterm.buffer.active.length ?? null;
-        return {
-          id: d.id,
-          cols: d.cols,
-          rows: d.rows,
-          renderer: d.renderer,
-          bufferLen,
-          scrollback: bufferLen !== null ? bufferLen - d.rows : null,
-          atlas: refs?.probes.webglAtlas() ?? null,
-          bufferBytes: refs?.probes.bufferBytes() ?? null,
-        };
-      }),
-      webgl,
-    };
+  const { snapshot, serverDiagnostics } = useDiagnosticSnapshot({
+    open: () => props.open,
+    activeId: () => props.activeId,
   });
-  const activeWatches = createMemo(
-    () =>
-      serverDiagnostics()?.trackedResources.filter(
-        (r) => r.kind === "fs-watch",
-      ) ?? [],
-  );
-  const activeResources = createMemo(
-    () =>
-      serverDiagnostics()?.trackedResources.filter(
-        (r) => r.kind !== "fs-watch",
-      ) ?? [],
-  );
 
   function copyJson() {
     void navigator.clipboard
@@ -179,433 +45,9 @@ const DiagnosticInfoContent: Component<{
       </div>
 
       <div class="overflow-y-auto">
-        <Section title="Browser" data-testid="diagnostic-browser">
-          <div class="space-y-0.5">
-            <Row label="Viewport">
-              <span class="text-fg">{isMobile() ? "mobile" : "canvas"}</span>
-            </Row>
-            <Row label="WebGL 2">
-              <span class={browser.webgl2Supported ? "text-ok" : "text-danger"}>
-                {browser.webgl2Supported ? "available" : "unavailable"}
-              </span>
-            </Row>
-            <Row label="DPR">
-              <span class="font-mono text-fg-3">
-                {browser.devicePixelRatio}
-              </span>
-            </Row>
-            <Show when={snapshot().session.jsHeap}>
-              {(heap) => (
-                <Row label="JS heap">
-                  <span class="font-mono text-fg">
-                    {heap().usedMB} / {heap().totalMB} MB
-                    <span class="text-fg-3/70"> (limit {heap().limitMB})</span>
-                  </span>
-                </Row>
-              )}
-            </Show>
-            <Row label="DOM">
-              <span class="font-mono text-fg">
-                {snapshot().session.domNodes}
-              </span>
-            </Row>
-            <Row label="Canvases">
-              <span class="font-mono text-fg">
-                {snapshot().session.canvases}
-              </span>
-            </Row>
-            <Row label="COI">
-              <span
-                class={browser.crossOriginIsolated ? "text-ok" : "text-fg-3"}
-              >
-                {browser.crossOriginIsolated ? "yes" : "no"}
-              </span>
-            </Row>
-            <Row label="UA">
-              <span class="font-mono text-fg-3 break-all">
-                {browser.userAgent}
-              </span>
-            </Row>
-          </div>
-        </Section>
-
-        <Section title="Xterm" data-testid="diagnostic-xterm">
-          <div class="space-y-0.5">
-            <Row label="xterm.js">
-              <span class="font-mono text-fg-3">{browser.xtermVersion}</span>
-            </Row>
-            <Row label="Active">
-              <span class="font-mono text-fg-3">
-                {props.activeId ? props.activeId.slice(0, 8) : "—"}
-              </span>
-            </Row>
-            <Row label="Count">
-              <span class="font-mono text-fg">{getDiagnostics().length}</span>
-            </Row>
-          </div>
-
-          <div class="mt-3 pt-2 border-t border-edge/50">
-            <div class="text-[10px] text-fg-3/70 mb-1">Terminals</div>
-          </div>
-          <Show
-            when={snapshot().terminals.length > 0}
-            fallback={
-              <div class="text-[11px] text-fg-3/60 italic">No terminals</div>
-            }
-          >
-            <div class="space-y-1">
-              <For each={snapshot().terminals}>
-                {(d) => (
-                  <div class="text-[11px] font-mono space-y-0.5">
-                    <div class="grid grid-cols-[9ch_8ch_1fr_auto] items-baseline gap-3">
-                      <span class="text-fg-3/70">{d.id.slice(0, 8)}</span>
-                      <span class="text-fg-2 tabular-nums">
-                        {d.cols}×{d.rows}
-                      </span>
-                      <span
-                        class={
-                          d.renderer === "webgl" ? "text-accent" : "text-fg-2"
-                        }
-                      >
-                        {d.renderer}
-                      </span>
-                      <Show when={props.activeId === d.id}>
-                        <span class="text-[10px] text-fg-3/70">active</span>
-                      </Show>
-                    </div>
-                    <Show when={d.scrollback !== null}>
-                      <div class="pl-[9ch] text-[10px] text-fg-3/60 tabular-nums">
-                        scrollback: {d.scrollback}
-                        <Show when={d.atlas}>
-                          {(a) => (
-                            <span>
-                              {" "}
-                              · atlas: {a().w}×{a().h}
-                            </span>
-                          )}
-                        </Show>
-                        <Show when={d.bufferBytes}>
-                          {(bb) => (
-                            <span>
-                              {" "}
-                              · buf: {formatMB(bb().primary)}
-                              <Show when={bb().alternate > 0}>
-                                {" "}
-                                (+alt {formatMB(bb().alternate)})
-                              </Show>
-                            </span>
-                          )}
-                        </Show>
-                      </div>
-                    </Show>
-                  </div>
-                )}
-              </For>
-            </div>
-          </Show>
-
-          <div class="mt-3 pt-2 border-t border-edge/50">
-            <div class="text-[10px] text-fg-3/70 mb-1">WebGL lifecycle</div>
-            <div class="space-y-0.5">
-              <Row label="Created">
-                <span class="font-mono text-fg tabular-nums">
-                  {snapshot().webgl.totalCreated}
-                </span>
-              </Row>
-              <Row label="Disposed">
-                <span class="font-mono text-fg tabular-nums">
-                  {snapshot().webgl.disposed}
-                </span>
-              </Row>
-              <Row label="In DOM">
-                <span class="font-mono text-fg tabular-nums">
-                  {snapshot().webgl.aliveInDom}
-                </span>
-              </Row>
-              <Row label="Zombies">
-                <span
-                  class={`font-mono tabular-nums ${
-                    snapshot().webgl.aliveDetached > 0
-                      ? "text-danger font-semibold"
-                      : "text-fg"
-                  }`}
-                >
-                  {snapshot().webgl.aliveDetached}
-                </span>
-              </Row>
-              <Row label="GCed">
-                <span class="font-mono text-fg-3 tabular-nums">
-                  {snapshot().webgl.gced}
-                </span>
-              </Row>
-              <Row label="Lost">
-                <span class="font-mono text-fg-3 tabular-nums">
-                  {snapshot().webgl.contextsLost}
-                </span>
-              </Row>
-            </div>
-          </div>
-
-          <Show when={snapshot().webgl.aliveCanvases.length > 0}>
-            <div class="mt-2 pt-2 border-t border-edge/50">
-              <div class="text-[10px] text-fg-3/70 mb-1">Alive canvases</div>
-              <div class="space-y-0.5 text-[10px] font-mono">
-                <For each={snapshot().webgl.aliveCanvases}>
-                  {(c) => (
-                    <div class="flex items-baseline gap-2 whitespace-nowrap">
-                      <span class="text-fg-3 tabular-nums w-[5ch] shrink-0">
-                        #{c.canvasId}
-                      </span>
-                      <span
-                        class={
-                          c.isConnected
-                            ? "text-fg-2 w-[9ch] shrink-0"
-                            : "text-danger w-[9ch] shrink-0"
-                        }
-                      >
-                        {c.isConnected ? "in-dom" : "detached"}
-                      </span>
-                      <span class="text-fg-2 tabular-nums">
-                        {c.width}×{c.height}
-                      </span>
-                      <span class="text-fg-3">·</span>
-                      <span class="text-fg-2 tabular-nums">
-                        {formatMB(c.bytesEst)}
-                      </span>
-                      <Show when={c.contextLost}>
-                        <span class="text-fg-3">·</span>
-                        <span class="text-fg-3/70">ctx-lost</span>
-                      </Show>
-                    </div>
-                  )}
-                </For>
-              </div>
-            </div>
-          </Show>
-
-          <details
-            data-testid="webgl-recent-events"
-            class="mt-2 pt-2 border-t border-edge/50 group"
-          >
-            <summary class="text-[10px] text-fg-3/70 cursor-pointer select-none list-none flex items-center justify-between gap-3">
-              <span>Recent events</span>
-              <span class="font-mono tabular-nums">
-                {snapshot().webgl.recentEvents.length}
-              </span>
-            </summary>
-            <Show
-              when={snapshot().webgl.recentEvents.length > 0}
-              fallback={
-                <div class="mt-1 text-[10px] text-fg-3/60 italic">
-                  No recent events
-                </div>
-              }
-            >
-              <div class="mt-1 space-y-0.5 text-[10px] font-mono overflow-x-auto">
-                <For each={snapshot().webgl.recentEvents}>
-                  {(ev) => (
-                    <div class="flex items-baseline gap-2 whitespace-nowrap">
-                      <span class="text-fg-3/60 tabular-nums shrink-0">
-                        {new Date(ev.ts).toISOString().slice(11, 23)}
-                      </span>
-                      <span class="text-fg-3 tabular-nums w-[5ch] shrink-0">
-                        #{ev.canvasId}
-                      </span>
-                      <span class="text-fg-2">
-                        {ev.kind}
-                        {ev.kind === "contextlost" && (
-                          <span class="text-fg-3/70">
-                            {" "}
-                            (defaultPrevented={String(ev.defaultPrevented)})
-                          </span>
-                        )}
-                      </span>
-                    </div>
-                  )}
-                </For>
-              </div>
-            </Show>
-          </details>
-        </Section>
-
-        <Section title="Server" data-testid="diagnostic-server">
-          <div class="space-y-0.5">
-            <Row label="WS" variant="badge">
-              {wsStatus()}
-            </Row>
-            <Show when={serverProcessId()}>
-              {(pid) => (
-                <Row label="Process">
-                  <span class="font-mono text-fg-3">{pid().slice(0, 8)}</span>
-                </Row>
-              )}
-            </Show>
-            <Show
-              when={serverDiagnostics()}
-              fallback={
-                <div
-                  data-testid="server-diagnostics-loading"
-                  class="text-[11px] text-fg-3/60 italic"
-                >
-                  {serverDiagnostics.error
-                    ? `Server diagnostics unavailable: ${serverDiagnostics.error.message}`
-                    : "Loading server diagnostics"}
-                </div>
-              }
-            >
-              {(server) => (
-                <>
-                  <Row label="Uptime">
-                    <span class="font-mono text-fg">
-                      {formatDuration(server().uptimeMs)}
-                    </span>
-                  </Row>
-                  <Row label="Memory">
-                    <span class="font-mono text-fg">
-                      rss {formatMB(server().memory.rss)}
-                      <span class="text-fg-3/70">
-                        {" "}
-                        · heap {formatMB(server().memory.heapUsed)} /{" "}
-                        {formatMB(server().memory.heapTotal)}
-                      </span>
-                    </span>
-                  </Row>
-                  <Row label="Publisher">
-                    <span class="font-mono text-fg">
-                      {server().counts.publisherSize}
-                    </span>
-                  </Row>
-                </>
-              )}
-            </Show>
-          </div>
-
-          <Show when={serverDiagnostics()}>
-            {(server) => (
-              <>
-                <div class="mt-3 pt-2 border-t border-edge/50">
-                  <div class="text-[10px] text-fg-3/70 mb-1">
-                    Terminal processes
-                  </div>
-                  <Show
-                    when={server().processes.length > 0}
-                    fallback={
-                      <div class="text-[11px] text-fg-3/60 italic">
-                        No terminal processes
-                      </div>
-                    }
-                  >
-                    <div class="space-y-1 text-[10px] font-mono">
-                      <For each={server().processes}>
-                        {(proc) => (
-                          <div class="grid grid-cols-[9ch_8ch_1fr_auto] items-baseline gap-3">
-                            <span class="text-fg-3/70">
-                              {proc.terminalId.slice(0, 8)}
-                            </span>
-                            <span class="text-fg-2 tabular-nums">
-                              pid {proc.pid}
-                            </span>
-                            <span class="text-fg-2 truncate" title={proc.cwd}>
-                              {proc.foregroundProcess ?? proc.cwd}
-                            </span>
-                            <Show when={proc.agentKind}>
-                              {(kind) => (
-                                <span class="text-accent">{kind()}</span>
-                              )}
-                            </Show>
-                          </div>
-                        )}
-                      </For>
-                    </div>
-                  </Show>
-                </div>
-
-                <div class="mt-3 pt-2 border-t border-edge/50">
-                  <div class="text-[10px] text-fg-3/70 mb-1">
-                    Active file system watches
-                  </div>
-                  <Show
-                    when={activeWatches().length > 0}
-                    fallback={
-                      <div class="text-[11px] text-fg-3/60 italic">
-                        No active file watches
-                      </div>
-                    }
-                  >
-                    <div class="space-y-1 text-[10px] font-mono">
-                      <For each={activeWatches()}>
-                        {(resource) => (
-                          <div>
-                            <div class="grid grid-cols-[12ch_14ch_1fr_6ch] items-baseline gap-3">
-                              <span class="text-fg-2">{resource.label}</span>
-                              <span class="text-fg-3/70">
-                                {resource.owner ?? "unknown"}
-                              </span>
-                              <span class="text-fg-2 break-all">
-                                {resource.target ?? "—"}
-                              </span>
-                              <span class="text-fg-3/70 tabular-nums">
-                                {formatDuration(
-                                  Math.max(
-                                    0,
-                                    server().sampledAt - resource.createdAt,
-                                  ),
-                                )}
-                              </span>
-                            </div>
-                            <Show when={formatDetails(resource.details)}>
-                              {(details) => (
-                                <div class="pl-[12ch] text-fg-3/60 break-all">
-                                  {details()}
-                                </div>
-                              )}
-                            </Show>
-                          </div>
-                        )}
-                      </For>
-                    </div>
-                  </Show>
-                </div>
-
-                <div class="mt-3 pt-2 border-t border-edge/50">
-                  <div class="text-[10px] text-fg-3/70 mb-1">
-                    Other tracked resources
-                  </div>
-                  <Show
-                    when={activeResources().length > 0}
-                    fallback={
-                      <div class="text-[11px] text-fg-3/60 italic">
-                        No active timers, subscriptions, or DB handles
-                      </div>
-                    }
-                  >
-                    <div class="space-y-1 text-[10px] font-mono">
-                      <For each={activeResources()}>
-                        {(resource) => (
-                          <div class="grid grid-cols-[12ch_13ch_1fr_6ch] items-baseline gap-3">
-                            <span class="text-fg-2">{resource.label}</span>
-                            <span class="text-fg-3/70">{resource.kind}</span>
-                            <span class="text-fg-2 break-all">
-                              {resource.target ?? resource.owner ?? "—"}
-                            </span>
-                            <span class="text-fg-3/70 tabular-nums">
-                              {formatDuration(
-                                Math.max(
-                                  0,
-                                  server().sampledAt - resource.createdAt,
-                                ),
-                              )}
-                            </span>
-                          </div>
-                        )}
-                      </For>
-                    </div>
-                  </Show>
-                </div>
-              </>
-            )}
-          </Show>
-        </Section>
+        <BrowserDiagnosticsSection snapshot={snapshot} />
+        <XtermDiagnosticsSection snapshot={snapshot} />
+        <ServerDiagnosticsSection serverDiagnostics={serverDiagnostics} />
       </div>
     </div>
   );
